// niokr-client/src/router/types/routeNames.ts
export const enum RouteNames {
  // auth
  LOGIN = 'LOGIN',

  // shared
  FAQ = 'FAQ',
  CONTACTS = 'CONTACTS',
  NOT_FOUND = 'NOT_FOUND',
  PROJECT_DETAILS = 'PROJECT_DETAILS',

  // user
  USER_PAGE = 'USER_PAGE',
  USER_INFO = 'USER_INFO',
  USER_PROJECTS = 'USER_PROJECTS',
  USER_EDIT_INFO = 'USER_EDIT_INFO',

  // customer
  CUSTOMER_PROJECT_CREATE = 'CUSTOMER_PROJECT_CREATE',

  // employee
  ACTIVE_PROJECTS = 'ACTIVE_PROJECTS',
  EMPLOYEE_APPLICATIONS = 'EMPLOYEE_APPLICATIONS',
  EMPLOYEE_HISTORY = 'EMPLOYEE_HISTORY',

  // admin
  ADMIN_TAGS = 'ADMIN_TAGS',

  // customer and admin
  ADD_EDIT_PROJECT = 'ADD_EDIT_PROJECT',
}